ref(scm): Move `should_increment_contributor_seat` to provider-agnostic location
We will need to re-use this logic for other SCM providers (e.g. GitLab soon), so moving this out of GitHub utils.
